Initial commit with base code idea & tests.
UNSTABLE CODE, PRONE TO CHANGES!

Base gem stub & codebase for salesforce interaction with
ActiveRecord.

There are a bunch of documentation blocks scattered through the
codebase, eventually they'll become more concrete (if they can
of course).

Base code is able to perform not so complex queries correctly,
although there is a lot of room for improvement, specially towards
SOQL specific syntax.

Goals for future commits, include tests that are able to reproduce
queries listed in [SFDC example queries] as a baseline for supported
features.

[SFDC example queries]: https://developer.salesforce.com/docs/atlas.en-us.soql_sosl.meta/soql_sosl/sforce_api_calls_soql_select_examples.htm